+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.9
+
+- add option `--dry-run` from `create`
+- better print when create
+
 ## 1.0.7
 
 - add option `--small-tx` for `create` to avoid error `Transaction too large`